+add_mapillary_tags seperate file

+++ python3/jpegoptimizer3.py
@@ -6,15 +6,14 @@
 #   purpose: optimize all JPG files with EXIF headers from a directory
 #            to 75% quality using OpenCV and Pillow before uploading
 #
+#   warning: deletes all broken pictures!
+#
 
 import os
 import sys
-# import cv2
-import numpy as np
 from tqdm import tqdm
 from PIL import Image
 from PIL import ImageOps
-# import piexif
 
 
 def optimize_file(file_path):
@@ -35,18 +34,15 @@ def optimize_file(file_path):
 #
 if __name__ == "__main__":
     print ("*** JPEG optimizer ***")
-    try:
-        ordnerpfad = sys.argv[1]
-    except:
-        ordnerpfad = r"D:\Mapillary\DCIM"
-        pass
+    ordnerpfad = sys.argv[1]
     # Loop over JPG files
     for subdirz, dirz, filez in os.walk(ordnerpfad):
-        print("\nOptimizing:", subdirz)
-        for f in tqdm(filez):
-            file_path = subdirz + os.sep + f
-            if file_path.lower().endswith('.jpg'):
-                success = optimize_file(file_path)
-                if not(success):
-                    os.remove(file_path)
+        if len(filez) > 0:
+            print("\nOptimizing:", subdirz)
+            for f in tqdm(filez):
+                file_path = subdirz + os.sep + f
+                if file_path.lower().endswith('.jpg'):
+                    success = optimize_file(file_path)
+                    if not(success):
+                        os.remove(file_path)
